Change type of _size fields to size_t

+++ src/rdkafka.h
@@ -2171,9 +2171,9 @@ RD_EXPORT
 void rd_kafka_conf_set_plain_creds_cb(rd_kafka_conf_t *conf,
                                       int (*plain_creds_cb)(rd_kafka_t *rk,
                                                             char *username,
-                                                            int *username_size,
+                                                            size_t *username_size,
                                                             char *password,
-                                                            int *password_size));
+                                                            size_t *password_size));
 
 /**
  * @brief Enable/disable creation of a queue specific to SASL events
